+++ example/.idea/misc.xml
@@ -5,10 +5,14 @@
       <map>
         <entry key="../../../../.gradle/caches/transforms-3/e94337de3ceb42b0135be1d8a4a9b7db/transformed/material-1.6.1/res/drawable/ic_clock_black_24dp.xml" value="0.4325" />
         <entry key="../../../../.gradle/caches/transforms-3/e94337de3ceb42b0135be1d8a4a9b7db/transformed/material-1.6.1/res/drawable/ic_m3_chip_check.xml" value="0.4855" />
+        <entry key="../../../../Library/Android/sdk/platforms/android-32/data/res/drawable-watch/sym_def_app_icon.xml" value="0.4855" />
+        <entry key="app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.4855" />
+        <entry key="app/src/main/res/drawable/ic_launcher_background.xml" value="0.4855" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.4217615763346355" />
         <entry key="app/src/main/res/layout/content_main.xml" value="0.384375" />
         <entry key="app/src/main/res/layout/fragment_first.xml" value="0.384375" />
         <entry key="app/src/main/res/menu/menu_main.xml" value="0.39010416666666664" />
+        <entry key="app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.4855" />
       </map>
     </option>
   </component>
