+++ example/android/build.gradle
@@ -3,13 +3,6 @@ plugins {
     id 'org.jetbrains.kotlin.android'
 }
 
-ext {
-    var = '12341234'
-    var1 = 'key0'
-    var2 = '12341234'
-    var3 = '/Users/mike/Desktop/android-release-key'
-}
-
 android {
     compileSdk 32
 
@@ -34,6 +27,7 @@ android {
     kotlinOptions {
         jvmTarget = '1.8'
     }
+
 }
 
 dependencies {
