+++ example/app/src/main/java/com/courier/example/ExampleApplication.kt
@@ -9,14 +9,14 @@ class ExampleApplication : Application() {
     override fun onCreate() {
         super.onCreate()
 
-        val options = FirebaseOptions.Builder().apply {
-            setApiKey("AIzaSyC_9Bq05Ywuy3mjAOkF8rB0LkmyYUoQIrA")
-            setApplicationId("1:694725526129:android:77c71528200b105c8811d0")
-            setProjectId("test-fcm-e7ddc")
-            setGcmSenderId("694725526129")
-        }.build()
-
-        FirebaseApp.initializeApp(this, options)
+//        val options = FirebaseOptions.Builder().apply {
+//            setApiKey("AIzaSyC_9Bq05Ywuy3mjAOkF8rB0LkmyYUoQIrA")
+//            setApplicationId("1:694725526129:android:77c71528200b105c8811d0")
+//            setProjectId("test-fcm-e7ddc")
+//            setGcmSenderId("694725526129")
+//        }.build()
+//
+//        FirebaseApp.initializeApp(this, options)
 
     }
